Accessible Metro Map
I apologize that it was not even a quarter done as expected. But I did
get a lot from this class. Before this, I had no idea of how front end
works and how codes structured. Now I have an overview of everything
and got this project that I am really passionate about. I submit this
idea to a NYU App design contest and made it in the semifinal. I will
keep working on it and try my best. Thank you for your help and
patience!